rename files and update axios instance
