+++ src/app/(root)/page.tsx
@@ -1,4 +1,4 @@
-import DiscoveryList from "@/components/home/DiscoveryList";
+import DiscoveryList from "@/components/home/discovery-list";
 
 export default async function Home() {
   return (
