+++ src/components/shadcn/Sidebar.tsx
@@ -7,23 +7,23 @@ import { PanelLeft } from "lucide-react";
 
 import { useIsMobile } from "@//hooks/useMobile";
 import { cn } from "@/lib/utils";
-import { Button } from "@/components/shadcn/Button";
-import { Input } from "@/components/shadcn/Input";
-import { Separator } from "@/components/shadcn/Separator";
+import { Button } from "@/components/shadcn/button";
+import { Input } from "@/components/shadcn/input";
+import { Separator } from "@/components/shadcn/separator";
 import {
   Sheet,
   SheetContent,
   SheetDescription,
   SheetHeader,
   SheetTitle,
-} from "@/components/shadcn/Sheet";
-import { Skeleton } from "@/components/shadcn/Skeleton";
+} from "@/components/shadcn/sheet";
+import { Skeleton } from "@/components/shadcn/skeleton";
 import {
   Tooltip,
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
-} from "@/components/shadcn/Tooltip";
+} from "@/components/shadcn/tooltip";
 
 const SIDEBAR_COOKIE_NAME = "sidebar_state";
 const SIDEBAR_COOKIE_MAX_AGE = 60 * 60 * 24 * 7;
