+++ src/components/shadcn/mode-toggle.tsx
@@ -4,13 +4,13 @@ import * as React from "react";
 import { Moon, MoonIcon, Sun } from "lucide-react";
 import { useTheme } from "next-themes";
 
-import { Button } from "./Button";
+import { Button } from "./button";
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
-} from "./DropdownMenu";
+} from "./dropdownMenu";
 
 export function ModeToggle() {
   const { setTheme } = useTheme();
