+++ src/services/band-service.ts
@@ -1,11 +1,11 @@
-import axios from "@/axios/axios";
 import Band from "@/interfaces/Band";
 import { ENDPOINTS } from "@/utils/constants";
+import httpClient from "./http-client";
 
 export const getBand = async (userId: string): Promise<Band> => {
-  return (await axios.get(`${ENDPOINTS.BAND}/user/${userId}`)).data;
+  return (await httpClient.get(`${ENDPOINTS.BAND}/user/${userId}`)).data;
 };
 
 export const updateBand = async (updateData: FormData) => {
-  await axios.put(`${ENDPOINTS.BANDUPDATE}`, updateData);
+  await httpClient.put(`${ENDPOINTS.BANDUPDATE}`, updateData);
 };
