+++ src/services/http-client.ts
@@ -1,7 +1,9 @@
 import { BASE_URL } from "@/utils/constants";
 import axios from "axios";
 
-export default axios.create({
+const httpClient = axios.create({
   baseURL: BASE_URL,
   withCredentials: true,
 });
+
+export default httpClient;
