+++ src/services/news-service.ts
