+++ src/services/spotify-service.ts
@@ -1,23 +1,25 @@
 import SearchSong from "@/interfaces/spotify/SearchSong";
-import axios from "../axios/axios";
 import SpotifyProfile from "../interfaces/spotify/SpotifyProfile";
 import { SPOTIFY_ENDPOINTS } from "@/utils/constants";
 import SpotifyArtist from "@/interfaces/spotify/SpotifyArtist";
+import httpClient from "./http-client";
 
 export const getSpotifyProfile = async (): Promise<SpotifyProfile> => {
-  const response = await axios.get(SPOTIFY_ENDPOINTS.PROFILE);
+  const response = await httpClient.get(SPOTIFY_ENDPOINTS.PROFILE);
   return response.data;
 };
 
 export const getSpotifySongsBySearch = async (
   search: string
 ): Promise<SearchSong[]> => {
-  const response = await axios.get(`${SPOTIFY_ENDPOINTS.SEARCH}/${search}`);
+  const response = await httpClient.get(
+    `${SPOTIFY_ENDPOINTS.SEARCH}/${search}`
+  );
   return response.data;
 };
 
 export const getSpotifyArtist = async (
   artistId: string | undefined
 ): Promise<SpotifyArtist> => {
-  return (await axios.get(`${SPOTIFY_ENDPOINTS.ARTIST}/${artistId}`)).data;
+  return (await httpClient.get(`${SPOTIFY_ENDPOINTS.ARTIST}/${artistId}`)).data;
 };
