+++ src/services/user-service.ts
@@ -1,6 +1,6 @@
 import UserSearchResult from "@/interfaces/user/UserSearchResult";
 import UserProfile from "../interfaces/UserProfile";
-import axios from "@/axios/axios";
+import httpClient from "./http-client";
 
 interface FollowData {
   loggedInUser: string;
@@ -10,19 +10,19 @@ interface FollowData {
 export const getUserProfile = async (
   username: string
 ): Promise<UserProfile> => {
-  const response = await axios.get(`/User/${username}/profile`);
+  const response = await httpClient.get(`/User/${username}/profile`);
   return response.data;
 };
 
 export const getUserByName = async (username: string) => {
-  const response = await axios.get(`User/${username}`);
+  const response = await httpClient.get(`User/${username}`);
   return response.data;
 };
 
 export const getUsersByNameSearch = async (
   username: string
 ): Promise<UserSearchResult[]> => {
-  const response = await axios.get(`/User/${username}`);
+  const response = await httpClient.get(`/User/${username}`);
   return response.data;
 };
 
@@ -30,7 +30,7 @@ export const checkFollowing = async (
   followerUsername?: string,
   followedUsername?: string
 ): Promise<boolean> => {
-  const response = await axios.get("/Follow/check-following", {
+  const response = await httpClient.get("/Follow/check-following", {
     params: {
       followerUsername,
       followedUsername,
@@ -41,7 +41,7 @@ export const checkFollowing = async (
 };
 
 export const followUser = async ({ loggedInUser, profileUser }: FollowData) => {
-  const response = await axios.post(`/Follow/create`, {
+  const response = await httpClient.post(`/Follow/create`, {
     loggedInUser,
     profileUser,
   });
@@ -53,7 +53,7 @@ export const unfollowUser = async ({
   loggedInUser,
   profileUser,
 }: FollowData) => {
-  const response = await axios.post(`/Follow/delete`, {
+  const response = await httpClient.post(`/Follow/delete`, {
     loggedInUser,
     profileUser,
   });
